fix(ext): address PR review comments
- Rename DEFAULT_FUNCTION_REGION to FUNCTION_DEFAULT_REGION in functionsEnvFromInstance.
- Use ref.version from parsed extensionRef if top-level extensionVersion is unset in ensureInstanceSpec.
- Co-locate ensureInstanceSpec unit tests in extensionsHelper.spec.ts.
