+++ src/extensions/extensionsHelper.spec.ts
@@ -12,6 +12,7 @@ import { storage } from "../gcp";
 import * as archiveDirectory from "../archiveDirectory";
 import * as prompt from "../prompt";
 import {
+  ExtensionInstance,
   ExtensionSource,
   ExtensionSpec,
   Param,
@@ -1000,6 +1001,208 @@ describe("extensionsHelper", () => {
           state: "PUBLISHED",
         }),
       ).to.eql("Prerelease");
+    });
+  });
+
+  describe("ensureInstanceSpec", () => {
+    let sandbox: sinon.SinonSandbox;
+
+    beforeEach(() => {
+      sandbox = sinon.createSandbox();
+    });
+
+    afterEach(() => {
+      sandbox.restore();
+    });
+
+    it("should return instance as is if spec already exists", async () => {
+      const instance: ExtensionInstance = {
+        name: "projects/123/instances/ext1",
+        createTime: "",
+        updateTime: "",
+        state: "ACTIVE",
+        serviceAccountEmail: "",
+        config: {
+          name: "",
+          createTime: "",
+          params: {},
+          systemParams: {},
+          source: {
+            name: "",
+            state: "ACTIVE",
+            packageUri: "",
+            hash: "",
+            spec: {
+              name: "my-ext",
+              version: "0.1.0",
+              resources: [],
+              params: [],
+              systemParams: [],
+            },
+          },
+        },
+      };
+
+      const getExtensionVersionStub = sandbox.stub(publisherApi, "getExtensionVersion");
+      const res = await extensionsHelper.ensureInstanceSpec(instance);
+      expect(res).to.equal(instance);
+      expect(getExtensionVersionStub).to.not.have.been.called;
+    });
+
+    it("should fetch spec on demand if missing", async () => {
+      const instance: ExtensionInstance = {
+        name: "projects/123/instances/ext1",
+        createTime: "",
+        updateTime: "",
+        state: "ACTIVE",
+        serviceAccountEmail: "",
+        config: {
+          name: "",
+          createTime: "",
+          params: {},
+          systemParams: {},
+          extensionRef: "firebase/firestore-send-email",
+          extensionVersion: "0.1.35",
+        },
+      };
+
+      sandbox.stub(publisherApi, "getExtensionVersion").resolves({
+        name: "publishers/firebase/extensions/firestore-send-email/versions/0.1.35",
+        ref: "firebase/firestore-send-email@0.1.35",
+        spec: {
+          name: "firestore-send-email",
+          version: "0.1.35",
+          resources: [],
+          params: [
+            {
+              param: "LOCATION",
+              label: "Location",
+              type: ParamType.SELECT,
+            },
+          ],
+          systemParams: [],
+        },
+        state: "PUBLISHED",
+        hash: "hash123",
+        sourceDownloadUri: "https://example.com/download",
+      });
+
+      const res = await extensionsHelper.ensureInstanceSpec(instance);
+      expect(res.config?.source?.spec?.name).to.equal("firestore-send-email");
+      expect(res.config?.source?.spec?.params).to.have.length(1);
+      expect(res.config?.source?.name).to.equal(
+        "publishers/firebase/extensions/firestore-send-email/versions/0.1.35",
+      );
+      expect(res.config?.source?.packageUri).to.equal("https://example.com/download");
+      expect(res.config?.source?.hash).to.equal("hash123");
+      expect(res.config?.source?.state).to.equal("ACTIVE");
+    });
+
+    it("should fetch spec on demand if extensionRef is on instance directly", async () => {
+      const instance: ExtensionInstance = {
+        name: "projects/123/instances/ext1",
+        createTime: "",
+        updateTime: "",
+        state: "ACTIVE",
+        serviceAccountEmail: "",
+        extensionRef: "firebase/firestore-send-email",
+        extensionVersion: "0.1.35",
+        config: {
+          name: "",
+          createTime: "",
+          params: {},
+          systemParams: {},
+        },
+      };
+
+      sandbox.stub(publisherApi, "getExtensionVersion").resolves({
+        name: "publishers/firebase/extensions/firestore-send-email/versions/0.1.35",
+        ref: "firebase/firestore-send-email@0.1.35",
+        spec: {
+          name: "firestore-send-email",
+          version: "0.1.35",
+          resources: [],
+          params: [
+            {
+              param: "LOCATION",
+              label: "Location",
+              type: ParamType.SELECT,
+            },
+          ],
+          systemParams: [],
+        },
+        state: "PUBLISHED",
+        hash: "hash123",
+        sourceDownloadUri: "https://example.com/download",
+      });
+
+      const res = await extensionsHelper.ensureInstanceSpec(instance);
+      expect(res.config?.source?.spec?.name).to.equal("firestore-send-email");
+      expect(res.config?.source?.spec?.params).to.have.length(1);
+      expect(res.config?.source?.name).to.equal(
+        "publishers/firebase/extensions/firestore-send-email/versions/0.1.35",
+      );
+    });
+
+    it("should preserve version from extensionRef when extensionVersion is unset", async () => {
+      const instance: ExtensionInstance = {
+        name: "projects/123/instances/ext1",
+        createTime: "",
+        updateTime: "",
+        state: "ACTIVE",
+        serviceAccountEmail: "",
+        extensionRef: "firebase/firestore-send-email@0.1.35",
+        config: {
+          name: "",
+          createTime: "",
+          params: {},
+          systemParams: {},
+        },
+      };
+
+      const getExtensionVersionStub = sandbox.stub(publisherApi, "getExtensionVersion").resolves({
+        name: "publishers/firebase/extensions/firestore-send-email/versions/0.1.35",
+        ref: "firebase/firestore-send-email@0.1.35",
+        spec: {
+          name: "firestore-send-email",
+          version: "0.1.35",
+          resources: [],
+          params: [],
+          systemParams: [],
+        },
+        state: "PUBLISHED",
+        hash: "hash123",
+        sourceDownloadUri: "https://example.com/download",
+      });
+
+      const res = await extensionsHelper.ensureInstanceSpec(instance);
+      expect(getExtensionVersionStub).to.have.been.calledWith(
+        "firebase/firestore-send-email@0.1.35",
+      );
+      expect(res.config?.source?.spec?.name).to.equal("firestore-send-email");
+    });
+
+    it("should let getExtensionVersion errors bubble up", async () => {
+      const instance: ExtensionInstance = {
+        name: "projects/123/instances/ext1",
+        createTime: "",
+        updateTime: "",
+        state: "ACTIVE",
+        serviceAccountEmail: "",
+        extensionRef: "firebase/firestore-send-email",
+        extensionVersion: "0.1.35",
+        config: {
+          name: "",
+          createTime: "",
+          params: {},
+          systemParams: {},
+        },
+      };
+
+      const networkErr = new Error("Network failure");
+      sandbox.stub(publisherApi, "getExtensionVersion").rejects(networkErr);
+
+      await expect(extensionsHelper.ensureInstanceSpec(instance)).to.be.rejectedWith(networkErr);
     });
   });
 });
